New Strategies folder added for Strategy class pattern.
