+++ src/Application/ygo.application/ApplicationInstaller.cs
@@ -23,7 +23,9 @@ using ygo.application.Queries.CardByName;
 using ygo.application.Queries.CategoryById;
 using ygo.application.Validations.Cards;
 using ygo.core.Services;
+using ygo.core.Strategies;
 using ygo.domain.Services;
+using ygo.domain.Strategies;
 
 namespace ygo.application
 {
