+++ src/Domain/ygo.domain/Services/CardService.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using ygo.core.Models;
 using ygo.core.Models.Db;
 using ygo.core.Services;
+using ygo.core.Strategies;
 using ygo.domain.Repository;
 
 namespace ygo.domain.Services
