+++ src/Domain/ygo.domain/Strategies/MonsterCardTypeStrategy.cs
@@ -1,44 +1,45 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using ygo.core.Enums;
 using ygo.core.Models;
 using ygo.core.Models.Db;
 using ygo.core.Services;
+using ygo.core.Strategies;
 using ygo.domain.Mappers;
 using ygo.domain.Repository;
 
-namespace ygo.domain.Services
+namespace ygo.domain.Strategies
 {
-    public class TrapCardTypeStrategy : ICardTypeStrategy
+    public class MonsterCardTypeStrategy : ICardTypeStrategy
     {
         private readonly ICardRepository _cardRepository;
 
-        public TrapCardTypeStrategy(ICardRepository cardRepository)
+        public MonsterCardTypeStrategy(ICardRepository cardRepository)
         {
             _cardRepository = cardRepository;
         }
-
         public async Task<Card> Add(CardModel cardModel)
         {
-            var newTrapCard = CardMapper.MapToSpellOrTrapCard(cardModel);
+            var newMonsterCard = CardMapper.MapToMonsterCard(cardModel);
 
-            return await _cardRepository.Add(newTrapCard);
+            return await _cardRepository.Add(newMonsterCard);
         }
+
         public async Task<Card> Update(CardModel cardModel)
         {
             var cardToUpdate = await _cardRepository.CardById(cardModel.Id);
 
             if (cardToUpdate != null)
             {
-                CardMapper.UpdateTrapCardWith(cardToUpdate, cardModel);
+                CardMapper.UpdateMonsterCardWith(cardToUpdate, cardModel);
                 return await _cardRepository.Update(cardToUpdate);
             }
 
             return null;
         }
+
         public bool Handles(YugiohCardType yugiohCardType)
         {
-            return yugiohCardType == YugiohCardType.Trap;
+            return yugiohCardType == YugiohCardType.Monster;
         }
     }
 }
