+++ src/Domain/ygo.domain/Strategies/TrapCardTypeStrategy.cs
@@ -1,45 +1,44 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using ygo.core.Enums;
 using ygo.core.Models;
 using ygo.core.Models.Db;
 using ygo.core.Services;
+using ygo.core.Strategies;
 using ygo.domain.Mappers;
 using ygo.domain.Repository;
 
-namespace ygo.domain.Services
+namespace ygo.domain.Strategies
 {
-    public class MonsterCardTypeStrategy : ICardTypeStrategy
+    public class TrapCardTypeStrategy : ICardTypeStrategy
     {
         private readonly ICardRepository _cardRepository;
 
-        public MonsterCardTypeStrategy(ICardRepository cardRepository)
+        public TrapCardTypeStrategy(ICardRepository cardRepository)
         {
             _cardRepository = cardRepository;
         }
+
         public async Task<Card> Add(CardModel cardModel)
         {
-            var newMonsterCard = CardMapper.MapToMonsterCard(cardModel);
+            var newTrapCard = CardMapper.MapToSpellOrTrapCard(cardModel);
 
-            return await _cardRepository.Add(newMonsterCard);
+            return await _cardRepository.Add(newTrapCard);
         }
-
         public async Task<Card> Update(CardModel cardModel)
         {
             var cardToUpdate = await _cardRepository.CardById(cardModel.Id);
 
             if (cardToUpdate != null)
             {
-                CardMapper.UpdateMonsterCardWith(cardToUpdate, cardModel);
+                CardMapper.UpdateTrapCardWith(cardToUpdate, cardModel);
                 return await _cardRepository.Update(cardToUpdate);
             }
 
             return null;
         }
-
         public bool Handles(YugiohCardType yugiohCardType)
         {
-            return yugiohCardType == YugiohCardType.Monster;
+            return yugiohCardType == YugiohCardType.Trap;
         }
     }
 }
